Move utf-8 utils to `utils`
